+++ data/build-info.js
@@ -1,6 +1,6 @@
 "use strict";
 window.CROSSROADS_BUILD_INFO = Object.freeze({
   engine: "Infantry Core",
-  version: "E1.4",
+  version: "E1.4.1",
   codename: "Procedural Terrain Containers"
 });